Fred51.py
Fred51.py started out as my compiled list of functions used to carry out
my daily work. It has evolved into a method that I can use to simply
automate sql scripts, as well as webscraping to allow the user to more
easily pull data off websites and store the data to be presented in
dashboards or automated emails.